Added tabbaritem chidlren to tab fragments
And added first fragment to test the idea works
