+++ NavigationReactNative/src/android/src/main/java/com/navigation/reactnative/TabBarManager.java
@@ -1,5 +1,7 @@
 package com.navigation.reactnative;
 
+import android.view.View;
+
 import androidx.annotation.NonNull;
 
 import com.facebook.react.uimanager.ThemedReactContext;
@@ -17,5 +19,25 @@ public class TabBarManager extends ViewGroupManager<TabBarView> {
     @Override
     protected TabBarView createViewInstance(@NonNull ThemedReactContext reactContext) {
         return new TabBarView(reactContext);
+    }
+
+    @Override
+    public int getChildCount(TabBarView parent) {
+        return parent.tabFragments.size();
+    }
+
+    @Override
+    public View getChildAt(TabBarView parent, int index) {
+        return parent.tabFragments.get(index).tabBarItem;
+    }
+
+    @Override
+    public void addView(TabBarView parent, View child, int index) {
+        parent.tabFragments.add(index, new TabFragment((TabBarItemView) child));
+    }
+
+    @Override
+    public void removeViewAt(TabBarView parent, int index) {
+        parent.tabFragments.remove(index);
     }
 }
